Extract tagging tests
