+++ tests/Container/ContainerTest.php
@@ -269,87 +269,6 @@ class ContainerTest extends TestCase
         $container->make(ContainerDependentStub::class, []);
     }
 
-    public function testContainerTags()
-    {
-        $container = new Container;
-        $container->tag(ContainerImplementationStub::class, 'foo', 'bar');
-        $container->tag(ContainerImplementationStubTwo::class, ['foo']);
-
-        $this->assertCount(1, $container->tagged('bar'));
-        $this->assertCount(2, $container->tagged('foo'));
-
-        $fooResults = [];
-        foreach ($container->tagged('foo') as $foo) {
-            $fooResults[] = $foo;
-        }
-
-        $barResults = [];
-        foreach ($container->tagged('bar') as $bar) {
-            $barResults[] = $bar;
-        }
-
-        $this->assertInstanceOf(ContainerImplementationStub::class, $fooResults[0]);
-        $this->assertInstanceOf(ContainerImplementationStub::class, $barResults[0]);
-        $this->assertInstanceOf(ContainerImplementationStubTwo::class, $fooResults[1]);
-
-        $container = new Container;
-        $container->tag([ContainerImplementationStub::class, ContainerImplementationStubTwo::class], ['foo']);
-        $this->assertCount(2, $container->tagged('foo'));
-
-        $fooResults = [];
-        foreach ($container->tagged('foo') as $foo) {
-            $fooResults[] = $foo;
-        }
-
-        $this->assertInstanceOf(ContainerImplementationStub::class, $fooResults[0]);
-        $this->assertInstanceOf(ContainerImplementationStubTwo::class, $fooResults[1]);
-
-        $this->assertCount(0, $container->tagged('this_tag_does_not_exist'));
-    }
-
-    public function testTaggedServicesAreLazyLoaded()
-    {
-        $container = $this->createPartialMock(Container::class, ['make']);
-        $container->expects($this->once())->method('make')->willReturn(new ContainerImplementationStub());
-
-        $container->tag(ContainerImplementationStub::class, ['foo']);
-        $container->tag(ContainerImplementationStubTwo::class, ['foo']);
-
-        $fooResults = [];
-        foreach ($container->tagged('foo') as $foo) {
-            $fooResults[] = $foo;
-            break;
-        }
-
-        $this->assertCount(2, $container->tagged('foo'));
-        $this->assertInstanceOf(ContainerImplementationStub::class, $fooResults[0]);
-    }
-
-    public function testLazyLoadedTaggedServicesCanBeLoopedOverMultipleTimes()
-    {
-        $container = new Container;
-        $container->tag(ContainerImplementationStub::class, 'foo');
-        $container->tag(ContainerImplementationStubTwo::class, ['foo']);
-
-        $services = $container->tagged('foo');
-
-        $fooResults = [];
-        foreach ($services as $foo) {
-            $fooResults[] = $foo;
-        }
-
-        $this->assertInstanceOf(ContainerImplementationStub::class, $fooResults[0]);
-        $this->assertInstanceOf(ContainerImplementationStubTwo::class, $fooResults[1]);
-
-        $fooResults = [];
-        foreach ($services as $foo) {
-            $fooResults[] = $foo;
-        }
-
-        $this->assertInstanceOf(ContainerImplementationStub::class, $fooResults[0]);
-        $this->assertInstanceOf(ContainerImplementationStubTwo::class, $fooResults[1]);
-    }
-
     public function testForgetInstanceForgetsInstance()
     {
         $container = new Container;
